refactor: Update import path for `getCustomConfig` in Conversation and Message models
